Move experiment-schema-related code to a new experiment module
