+++ pyquil/operator_estimation.py
@@ -1,423 +1,25 @@
 import functools
 import itertools
-import json
 import logging
-import re
-import sys
 import warnings
-from json import JSONEncoder
+from math import pi
 from operator import mul
-from typing import List, Union, Iterable, Tuple, Optional, Dict, Callable
+from typing import Callable, Dict, List, Union, Iterable, Tuple, Optional
 
 import networkx as nx
 import numpy as np
 from networkx.algorithms.approximation.clique import clique_removal
-from functools import reduce
+
 from pyquil import Program
 from pyquil.api import QuantumComputer
+from pyquil.experiment import (_OneQState, _pauli_to_product_state, ExperimentResult,
+                               ExperimentSetting, SymmetrizationLevel, TensorProductState,
+                               TomographyExperiment)
 from pyquil.gates import *
 from pyquil.paulis import PauliTerm, sI, is_identity
-from math import pi
-from enum import IntEnum
 
-if sys.version_info < (3, 7):
-    from pyquil.external.dataclasses import dataclass
-else:
-    from dataclasses import dataclass
 
 log = logging.getLogger(__name__)
-
-
-class SymmetrizationLevel(IntEnum):
-    EXHAUSTIVE = -1
-    NONE = 0
-    OA_STRENGTH_1 = 1
-    OA_STRENGTH_2 = 2
-    OA_STRENGTH_3 = 3
-
-
-@dataclass(frozen=True)
-class _OneQState:
-    """
-    A description of a named one-qubit quantum state.
-
-    This can be used to generate pre-rotations for quantum process tomography. For example,
-    X0_14 will generate the +1 eigenstate of the X operator on qubit 14. X1_14 will generate the
-    -1 eigenstate. SIC0_14 will generate the 0th SIC-basis state on qubit 14.
-    """
-    label: str
-    index: int
-    qubit: int
-
-    def __str__(self):
-        return f'{self.label}{self.index}_{self.qubit}'
-
-    @classmethod
-    def from_str(cls, s):
-        ma = re.match(r'\s*(\w+)(\d+)_(\d+)\s*', s)
-        if ma is None:
-            raise ValueError(f"Couldn't parse '{s}'")
-        return _OneQState(
-            label=ma.group(1),
-            index=int(ma.group(2)),
-            qubit=int(ma.group(3)),
-        )
-
-
-@dataclass(frozen=True)
-class TensorProductState:
-    """
-    A description of a multi-qubit quantum state that is a tensor product of many _OneQStates
-    states.
-    """
-    states: Tuple[_OneQState]
-
-    def __init__(self, states=None):
-        if states is None:
-            states = tuple()
-        object.__setattr__(self, 'states', tuple(states))
-
-    def __mul__(self, other):
-        return TensorProductState(self.states + other.states)
-
-    def __str__(self):
-        return ' * '.join(str(s) for s in self.states)
-
-    def __repr__(self):
-        return f'TensorProductState[{self}]'
-
-    def __getitem__(self, qubit):
-        """Return the _OneQState at the given qubit."""
-        for oneq_state in self.states:
-            if oneq_state.qubit == qubit:
-                return oneq_state
-        raise IndexError()
-
-    def __iter__(self):
-        yield from self.states
-
-    def __len__(self):
-        return len(self.states)
-
-    def states_as_set(self):
-        return frozenset(self.states)
-
-    def __eq__(self, other):
-        if not isinstance(other, TensorProductState):
-            return False
-
-        return self.states_as_set() == other.states_as_set()
-
-    def __hash__(self):
-        return hash(self.states_as_set())
-
-    @classmethod
-    def from_str(cls, s):
-        if s == '':
-            return TensorProductState()
-        return TensorProductState(tuple(_OneQState.from_str(x) for x in s.split('*')))
-
-
-def _pauli_to_product_state(in_state: PauliTerm) -> TensorProductState:
-    """
-    Convert a Pauli term to a TensorProductState.
-    """
-    if is_identity(in_state):
-        in_state = TensorProductState()
-    else:
-        in_state = TensorProductState([
-            _OneQState(label=pauli_label, index=0, qubit=qubit)
-            for qubit, pauli_label in in_state._ops.items()
-        ])
-    return in_state
-
-
-def SIC0(q):
-    return TensorProductState((_OneQState('SIC', 0, q),))
-
-
-def SIC1(q):
-    return TensorProductState((_OneQState('SIC', 1, q),))
-
-
-def SIC2(q):
-    return TensorProductState((_OneQState('SIC', 2, q),))
-
-
-def SIC3(q):
-    return TensorProductState((_OneQState('SIC', 3, q),))
-
-
-def plusX(q):
-    return TensorProductState((_OneQState('X', 0, q),))
-
-
-def minusX(q):
-    return TensorProductState((_OneQState('X', 1, q),))
-
-
-def plusY(q):
-    return TensorProductState((_OneQState('Y', 0, q),))
-
-
-def minusY(q):
-    return TensorProductState((_OneQState('Y', 1, q),))
-
-
-def plusZ(q):
-    return TensorProductState((_OneQState('Z', 0, q),))
-
-
-def minusZ(q):
-    return TensorProductState((_OneQState('Z', 1, q),))
-
-
-def zeros_state(qubits: Iterable[int]):
-    return TensorProductState(_OneQState('Z', 0, q) for q in qubits)
-
-
-@dataclass(frozen=True, init=False)
-class ExperimentSetting:
-    """
-    Input and output settings for a tomography-like experiment.
-
-    Many near-term quantum algorithms take the following form:
-
-     - Start in a pauli state
-     - Prepare some ansatz
-     - Measure it w.r.t. pauli operators
-
-    Where we typically use a large number of (start, measure) pairs but keep the ansatz preparation
-    program consistent. This class represents the (start, measure) pairs. Typically a large
-    number of these :py:class:`ExperimentSetting` objects will be created and grouped into
-    a :py:class:`TomographyExperiment`.
-    """
-    in_state: TensorProductState
-    out_operator: PauliTerm
-
-    def __init__(self, in_state: TensorProductState, out_operator: PauliTerm):
-        # For backwards compatibility, handle in_state specified by PauliTerm.
-        if isinstance(in_state, PauliTerm):
-            warnings.warn("Please specify in_state as a TensorProductState",
-                          DeprecationWarning, stacklevel=2)
-            in_state = _pauli_to_product_state(in_state)
-        object.__setattr__(self, 'in_state', in_state)
-        object.__setattr__(self, 'out_operator', out_operator)
-
-    @property
-    def in_operator(self):
-        warnings.warn("ExperimentSetting.in_operator is deprecated in favor of in_state",
-                      stacklevel=2)
-
-        # Backwards compat
-        pt = sI()
-        for oneq_state in self.in_state.states:
-            if oneq_state.label not in ['X', 'Y', 'Z']:
-                raise ValueError(f"Can't shim {oneq_state.label} into a pauli term. Use in_state.")
-            if oneq_state.index != 0:
-                raise ValueError(f"Can't shim {oneq_state} into a pauli term. Use in_state.")
-
-            pt *= PauliTerm(op=oneq_state.label, index=oneq_state.qubit)
-
-        return pt
-
-    def __str__(self):
-        return f'{self.in_state}→{self.out_operator.compact_str()}'
-
-    def __repr__(self):
-        return f'ExperimentSetting[{self}]'
-
-    def serializable(self):
-        return str(self)
-
-    @classmethod
-    def from_str(cls, s: str):
-        """The opposite of str(expt)"""
-        instr, outstr = s.split('→')
-        return ExperimentSetting(in_state=TensorProductState.from_str(instr),
-                                 out_operator=PauliTerm.from_compact_str(outstr))
-
-
-def _abbrev_program(program: Program, max_len=10):
-    """Create an abbreviated string representation of a Program.
-
-    This will join all instructions onto a single line joined by '; '. If the number of
-    instructions exceeds ``max_len``, some will be excluded from the string representation.
-    """
-    program_lines = program.out().splitlines()
-    if max_len is not None and len(program_lines) > max_len:
-        first_n = max_len // 2
-        last_n = max_len - first_n
-        excluded = len(program_lines) - max_len
-        program_lines = (program_lines[:first_n] + [f'... {excluded} instrs not shown ...']
-                         + program_lines[-last_n:])
-
-    return '; '.join(program_lines)
-
-
-class TomographyExperiment:
-    """
-    A tomography-like experiment.
-
-    Many near-term quantum algorithms involve:
-
-     - some limited state preparation
-     - enacting a quantum process (like in tomography) or preparing a variational ansatz state
-       (like in VQE)
-     - measuring observables of the state.
-
-    Where we typically use a large number of (state_prep, measure) pairs but keep the ansatz
-    program consistent. This class stores the ansatz program as a :py:class:`~pyquil.Program`
-    and maintains a list of :py:class:`ExperimentSetting` objects which each represent a
-    (state_prep, measure) pair.
-
-    Settings diagonalized by a shared tensor product basis (TPB) can (optionally) be estimated
-    simultaneously. Therefore, this class is backed by a list of list of ExperimentSettings.
-    Settings sharing an inner list will be estimated simultaneously. If you don't want this,
-    provide a list of length-1-lists. As a convenience, if you pass a 1D list to the constructor
-    will expand it to a list of length-1-lists.
-
-    This class will not group settings for you. Please see :py:func:`group_experiments` for
-    a function that will automatically process a TomographyExperiment to group Experiments sharing
-    a TPB.
-    """
-
-    def __init__(self,
-                 settings: Union[List[ExperimentSetting], List[List[ExperimentSetting]]],
-                 program: Program,
-                 qubits: List[int] = None):
-        if len(settings) == 0:
-            settings = []
-        else:
-            if isinstance(settings[0], ExperimentSetting):
-                # convenience wrapping in lists of length 1
-                settings = [[expt] for expt in settings]
-
-        self._settings = settings  # type: List[List[ExperimentSetting]]
-        self.program = program
-        if qubits is not None:
-            warnings.warn("The 'qubits' parameter has been deprecated and will be removed"
-                          "in a future release of pyquil")
-        self.qubits = qubits
-
-    def __len__(self):
-        return len(self._settings)
-
-    def __getitem__(self, item):
-        return self._settings[item]
-
-    def __setitem__(self, key, value):
-        self._settings[key] = value
-
-    def __delitem__(self, key):
-        self._settings.__delitem__(key)
-
-    def __iter__(self):
-        yield from self._settings
-
-    def __reversed__(self):
-        yield from reversed(self._settings)
-
-    def __contains__(self, item):
-        return item in self._settings
-
-    def append(self, expts):
-        if not isinstance(expts, list):
-            expts = [expts]
-        return self._settings.append(expts)
-
-    def count(self, expt):
-        return self._settings.count(expt)
-
-    def index(self, expt, start=None, stop=None):
-        return self._settings.index(expt, start, stop)
-
-    def extend(self, expts):
-        return self._settings.extend(expts)
-
-    def insert(self, index, expt):
-        return self._settings.insert(index, expt)
-
-    def pop(self, index=None):
-        return self._settings.pop(index)
-
-    def remove(self, expt):
-        return self._settings.remove(expt)
-
-    def reverse(self):
-        return self._settings.reverse()
-
-    def sort(self, key=None, reverse=False):
-        return self._settings.sort(key, reverse)
-
-    def setting_strings(self):
-        yield from ('{i}: {st_str}'.format(i=i, st_str=', '.join(str(setting)
-                                                                 for setting in settings))
-                    for i, settings in enumerate(self._settings))
-
-    def settings_string(self, abbrev_after=None):
-        setting_strs = list(self.setting_strings())
-        if abbrev_after is not None and len(setting_strs) > abbrev_after:
-            first_n = abbrev_after // 2
-            last_n = abbrev_after - first_n
-            excluded = len(setting_strs) - abbrev_after
-            setting_strs = (setting_strs[:first_n] + [f'... {excluded} not shown ...',
-                                                      '... use e.settings_string() for all ...']
-                            + setting_strs[-last_n:])
-        return '\n'.join(setting_strs)
-
-    def __str__(self):
-        return _abbrev_program(self.program) + '\n' + self.settings_string(abbrev_after=20)
-
-    def serializable(self):
-        return {
-            'type': 'TomographyExperiment',
-            'settings': self._settings,
-            'program': self.program.out(),
-        }
-
-    def __eq__(self, other):
-        if not isinstance(other, TomographyExperiment):
-            return False
-        return self.serializable() == other.serializable()
-
-
-class OperatorEncoder(JSONEncoder):
-    def default(self, o):
-        if isinstance(o, ExperimentSetting):
-            return o.serializable()
-        if isinstance(o, TomographyExperiment):
-            return o.serializable()
-        if isinstance(o, ExperimentResult):
-            return o.serializable()
-        return o
-
-
-def to_json(fn, obj):
-    """Convenience method to save pyquil.operator_estimation objects as a JSON file.
-
-    See :py:func:`read_json`.
-    """
-    with open(fn, 'w') as f:
-        json.dump(obj, f, cls=OperatorEncoder, indent=2, ensure_ascii=False)
-    return fn
-
-
-def _operator_object_hook(obj):
-    if 'type' in obj and obj['type'] == 'TomographyExperiment':
-        return TomographyExperiment([[ExperimentSetting.from_str(s) for s in settings]
-                                     for settings in obj['settings']],
-                                    program=Program(obj['program']))
-    return obj
-
-
-def read_json(fn):
-    """Convenience method to read pyquil.operator_estimation objects from a JSON file.
-
-    See :py:func:`to_json`.
-    """
-    with open(fn) as f:
-        return json.load(f, object_hook=_operator_object_hook)
 
 
 def _one_q_sic_prep(index, qubit):
@@ -715,112 +317,6 @@ def group_experiments(experiments: TomographyExperiment,
         return group_experiments_greedy(experiments)
     elif method == 'clique-removal':
         return group_experiments_clique_removal(experiments)
-
-
-@dataclass(frozen=True)
-class ExperimentResult:
-    """An expectation and standard deviation for the measurement of one experiment setting
-    in a tomographic experiment.
-
-    In the case of readout error calibration, we also include
-    expectation, standard deviation and count for the calibration results, as well as the
-    expectation and standard deviation for the corrected results.
-    """
-
-    setting: ExperimentSetting
-    expectation: Union[float, complex]
-    total_counts: int
-    std_err: Union[float, complex] = None
-    raw_expectation: Union[float, complex] = None
-    raw_std_err: float = None
-    calibration_expectation: Union[float, complex] = None
-    calibration_std_err: Union[float, complex] = None
-    calibration_counts: int = None
-
-    def __init__(self, setting: ExperimentSetting,
-                 expectation: Union[float, complex],
-                 total_counts: int,
-                 stddev: Union[float, complex] = None,
-                 std_err: Union[float, complex] = None,
-                 raw_expectation: Union[float, complex] = None,
-                 raw_stddev: float = None,
-                 raw_std_err: float = None,
-                 calibration_expectation: Union[float, complex] = None,
-                 calibration_stddev: Union[float, complex] = None,
-                 calibration_std_err: Union[float, complex] = None,
-                 calibration_counts: int = None):
-
-        object.__setattr__(self, 'setting', setting)
-        object.__setattr__(self, 'expectation', expectation)
-        object.__setattr__(self, 'total_counts', total_counts)
-        object.__setattr__(self, 'raw_expectation', raw_expectation)
-        object.__setattr__(self, 'calibration_expectation', calibration_expectation)
-        object.__setattr__(self, 'calibration_counts', calibration_counts)
-
-        if stddev is not None:
-            warnings.warn("'stddev' has been renamed to 'std_err'")
-            std_err = stddev
-        object.__setattr__(self, 'std_err', std_err)
-
-        if raw_stddev is not None:
-            warnings.warn("'raw_stddev' has been renamed to 'raw_std_err'")
-            raw_std_err = raw_stddev
-        object.__setattr__(self, 'raw_std_err', raw_std_err)
-
-        if calibration_stddev is not None:
-            warnings.warn("'calibration_stddev' has been renamed to 'calibration_std_err'")
-            calibration_std_err = calibration_stddev
-        object.__setattr__(self, 'calibration_std_err', calibration_std_err)
-
-    def get_stddev(self) -> Union[float, complex]:
-        warnings.warn("'stddev' has been renamed to 'std_err'")
-        return self.std_err
-
-    def set_stddev(self, value: Union[float, complex]):
-        warnings.warn("'stddev' has been renamed to 'std_err'")
-        object.__setattr__(self, 'std_err', value)
-
-    stddev = property(get_stddev, set_stddev)
-
-    def get_raw_stddev(self) -> float:
-        warnings.warn("'raw_stddev' has been renamed to 'raw_std_err'")
-        return self.raw_std_err
-
-    def set_raw_stddev(self, value: float):
-        warnings.warn("'raw_stddev' has been renamed to 'raw_std_err'")
-        object.__setattr__(self, 'raw_std_err', value)
-
-    raw_stddev = property(get_raw_stddev, set_raw_stddev)
-
-    def get_calibration_stddev(self) -> Union[float, complex]:
-        warnings.warn("'calibration_stddev' has been renamed to 'calibration_std_err'")
-        return self.calibration_std_err
-
-    def set_calibration_stddev(self, value: Union[float, complex]):
-        warnings.warn("'calibration_stddev' has been renamed to 'calibration_std_err'")
-        object.__setattr__(self, 'calibration_std_err', value)
-
-    calibration_stddev = property(get_calibration_stddev, set_calibration_stddev)
-
-    def __str__(self):
-        return f'{self.setting}: {self.expectation} +- {self.std_err}'
-
-    def __repr__(self):
-        return f'ExperimentResult[{self}]'
-
-    def serializable(self):
-        return {
-            'type': 'ExperimentResult',
-            'setting': self.setting,
-            'expectation': self.expectation,
-            'std_err': self.std_err,
-            'total_counts': self.total_counts,
-            'raw_expectation': self.raw_expectation,
-            'raw_std_err': self.raw_std_err,
-            'calibration_expectation': self.calibration_expectation,
-            'calibration_std_err': self.calibration_std_err,
-            'calibration_counts': self.calibration_counts,
-        }
 
 
 def _generate_experiment_programs(
